Avoid downloading entire segments to memory
Currently chunks are downloaded to memory and written to disk.
When you are downloading a large number of HD m3u8 streamins in
parallel, this can lead to a memory crunch. By using file objects
and making use of AES cipher's ability to encrypt/decrypt in chunks,
we can reduce this memory utilization